+++ packages/core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.10.0 - 2019-05-23
+
+#### 🚀 Updates
+
+- **[Chip]** Add new compact and active props (#66) ([aefd828](https://github.com/airbnb/lunar/commit/aefd828)), closes [#66](https://github.com/airbnb/lunar/issues/66)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+### 1.9.1 - 2019-05-21
+
+#### 🐞 Fixes
+
+- Remove Babel runtime setup (#65) ([a58ade2](https://github.com/airbnb/lunar/commit/a58ade2)), closes [#65](https://github.com/airbnb/lunar/issues/65)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
 ## 1.9.0 - 2019-05-16
 
 #### 🚀 Updates